requirements.txt
numpy==1.20.2
Pillow==8.2.0
pkg-resources==0.0.0
pytz==2021.1
RPi.GPIO==0.7.0
spidev==3.5
